Move helper functions to shared folder
Helper functions have been moved to a new folder within the root directory.
This makes them easier to access between rest and graphql
