+++ keys/apiKeyAuth.js
@@ -1,7 +1,7 @@
 const faunadb = require('faunadb');
 const { keys } = require('underscore');
 const crypto = require('crypto')
-var {createErrorJSON} = require("../rest/v0/errors.helper")
+var {createErrorJSON} = require("../helpers/errors.helper")
 const client = new faunadb.Client({ secret: process.env.FAUNADB_KEY});
 const {
     Get,
