+++ rest/v0/grades.js
@@ -1,6 +1,6 @@
 var express = require("express");
 var router = express.Router();
-var {parseGradesParamsToSQL, queryDatabaseAndResponse} = require('./grades.helper')
+var {parseGradesParamsToSQL, queryDatabaseAndResponse} = require('../../helpers/grades.helper')
 
 // var { apiKeyAuth } = require("../../keys/apiKeyAuth");
